Attempt #1 improving quality
code is producing sharper images but it produces first an image that is the correct pixel ratio to 688 but isn't resized. Resizing it manually does come out well though. The second picture it produces is the normal pixelated 688 photos.

need to rework code so that it resizes the sharpened photos down to 688 and 320